test: adding tests for coverage
